Version 1.76.3
Fixed an issue which caused a spam of "Could not pass event InventoryCreativeEvent to ServerEssentials" this is no longer the case.
You can now set someone into Fly mode via the Console
You can now TabComplete the SendHome, SendWarp and DeleteWarp commands in Console
Went through each and every Command so that now when the Command isn't supported for Console Use, instead of posting an Error it will simply post the new Lang.yml Variable 'console-invalid'.

+++ src/main/java/me/rocketmankianproductions/serveressentials/commands/DeleteSpawn.java
@@ -1,6 +1,7 @@
 package me.rocketmankianproductions.serveressentials.commands;
 
 import me.rocketmankianproductions.serveressentials.file.Lang;
+import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
@@ -11,25 +12,31 @@ import org.jetbrains.annotations.NotNull;
 public class DeleteSpawn implements CommandExecutor {
     @Override
     public boolean onCommand(@NotNull CommandSender sender, @NotNull Command command, @NotNull String label, @NotNull String[] args) {
-        Player player = (Player) sender;
-        // Checking if the player has the se.deletespawn permission
-        if (player.hasPermission("se.deletespawn") || player.hasPermission("se.all")) {
-            // Averaging out the whether the file exists or not by checking for value in one of the default saving points
-            if (Setspawn.file.exists() && Setspawn.fileConfig.getString("Location.World") != null) {
-                // If the file exists then it will get deleted upon execution of command
-                Setspawn.file.delete();
-                Setspawn.reload();
-                String msg = Lang.fileConfig.getString("spawn-deletion-success");
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', msg));
-                return true;
+        if (sender instanceof Player){
+            Player player = (Player) sender;
+            // Checking if the player has the se.deletespawn permission
+            if (player.hasPermission("se.deletespawn") || player.hasPermission("se.all")) {
+                // Averaging out the whether the file exists or not by checking for value in one of the default saving points
+                if (Setspawn.file.exists() && Setspawn.fileConfig.getString("Location.World") != null) {
+                    // If the file exists then it will get deleted upon execution of command
+                    Setspawn.file.delete();
+                    Setspawn.reload();
+                    String msg = Lang.fileConfig.getString("spawn-deletion-success");
+                    player.sendMessage(ChatColor.translateAlternateColorCodes('&', msg));
+                    return true;
+                } else {
+                    String msg = Lang.fileConfig.getString("spawn-not-found");
+                    player.sendMessage(ChatColor.translateAlternateColorCodes('&', msg));
+                    return true;
+                }
             } else {
-                String msg = Lang.fileConfig.getString("spawn-not-found");
-                player.sendMessage(ChatColor.translateAlternateColorCodes('&', msg));
+                String perm = Lang.fileConfig.getString("no-permission-message").replace("<permission>", "se.deletespawn");
+                player.sendMessage(ChatColor.translateAlternateColorCodes('&', perm));
                 return true;
             }
-        } else {
-            String perm = Lang.fileConfig.getString("no-permission-message").replace("<permission>", "se.deletespawn");
-            player.sendMessage(ChatColor.translateAlternateColorCodes('&', perm));
+        }else{
+            String console = Lang.fileConfig.getString("console-invalid");
+            Bukkit.getLogger().info(ChatColor.translateAlternateColorCodes('&', console));
             return true;
         }
     }
